tests: Parametrize apps name in benchmark
