+++ python/tox.ini
@@ -78,4 +78,4 @@ pass_env =
     CI
 commands =
     poetry install --with integration,benchmark
-    poetry run pytest -vv --tb native --log-cli-level=INFO -s {[vars]tests_path}/integration/test_benchmark_kyuubi.py {posargs}
+    poetry run pytest -vv --tb native --log-cli-level=INFO -s {[vars]tests_path}/integration/bench/test_benchmark_kyuubi.py {posargs}
